feat(online-evals): feature-complete code evaluator details for projects
Bring the project code evaluator details page to parity with the dataset
code evaluator details:

- Sandbox, annotation configuration, and input mapping cards in the
  overview aside for code evaluators
- Versions tab with author, relative time, and diff-against-previous
- Remove the overflowing language field from the Activity panel (the
  language still shows on the Source Code card)

Shared components extracted into components/evaluators/ and reused by
both the dataset and project pages: CodeEvaluatorSandboxCard (fetches
sandboxBackends itself so only pages that render it pay for the backend
probes), EvaluatorAnnotationsCard, EvaluatorInputMappingCard, and
CodeEvaluatorVersions (queries the CodeEvaluator node directly).
