+++ js/app/src/pages/dataset/evaluators/CodeDatasetEvaluatorDetails.tsx
@@ -1,236 +1,27 @@
-import { css } from "@emotion/react";
-import { useMemo } from "react";
 import { useFragment } from "react-relay";
 import { graphql } from "relay-runtime";
 import invariant from "tiny-invariant";
 
-import {
-  Card,
-  ContextualHelp,
-  Empty,
-  Flex,
-  Icon,
-  Icons,
-  LinkButton,
-  Text,
-  View,
-} from "@phoenix/components";
+import { Card, Empty, Flex } from "@phoenix/components";
+import { CodeEvaluatorSandboxCard } from "@phoenix/components/evaluators/CodeEvaluatorSandboxCard";
 import { CodeEvaluatorSourceCodeBlock } from "@phoenix/components/evaluators/CodeEvaluatorSourceCodeBlock";
+import { EvaluatorAnnotationsCard } from "@phoenix/components/evaluators/EvaluatorAnnotationsCard";
 import {
-  EvaluatorDetailList,
-  EvaluatorDetailRow,
   evaluatorSplitContainerCSS,
   evaluatorSplitLayoutCSS,
 } from "@phoenix/components/evaluators/EvaluatorDetailsSection";
-import type { OutputConfig } from "@phoenix/components/evaluators/OutputConfigBlock";
-import { OutputConfigBlock } from "@phoenix/components/evaluators/OutputConfigBlock";
-import { SandboxProviderIcon } from "@phoenix/components/sandbox/SandboxProviderIcon";
-import { useViewerCanManageSandboxes } from "@phoenix/contexts";
+import { EvaluatorInputMappingCard } from "@phoenix/components/evaluators/EvaluatorInputMappingCard";
 import type { CodeDatasetEvaluatorDetails_datasetEvaluator$key } from "@phoenix/pages/dataset/evaluators/__generated__/CodeDatasetEvaluatorDetails_datasetEvaluator.graphql";
-import type { datasetEvaluatorDetailsLoaderQuery } from "@phoenix/pages/dataset/evaluators/__generated__/datasetEvaluatorDetailsLoaderQuery.graphql";
-import {
-  getSandboxConfigSettings,
-  LanguageWithIcon,
-} from "@phoenix/pages/settings/sandboxes/utils";
-
-type SandboxBackendInfo =
-  datasetEvaluatorDetailsLoaderQuery["response"]["sandboxBackends"][number];
-
-const mapGridCSS = css`
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: var(--global-dimension-size-150);
-
-  @media (max-width: 1100px) {
-    grid-template-columns: 1fr 1fr;
-  }
-
-  @media (max-width: 720px) {
-    grid-template-columns: 1fr;
-  }
-`;
-
-function CapabilityRow({ label, value }: { label: string; value: string }) {
-  return (
-    <Flex direction="row" gap="size-200" justifyContent="space-between">
-      <Text size="XS" color="text-700">
-        {label}
-      </Text>
-      <Text size="XS">{value}</Text>
-    </Flex>
-  );
-}
-
-function ProviderCapabilitiesHelp({
-  sandboxBackend,
-}: {
-  sandboxBackend: SandboxBackendInfo | undefined;
-}) {
-  return (
-    <ContextualHelp variant="info">
-      <Flex direction="column" gap="size-100">
-        <Text weight="heavy" size="S">
-          Capabilities
-        </Text>
-        <Flex direction="column" gap="size-50">
-          <CapabilityRow
-            label="env_vars"
-            value={
-              sandboxBackend?.supportsEnvVars ? "supported" : "not supported"
-            }
-          />
-          <CapabilityRow
-            label="internet_access"
-            value={getInternetAccessLabel(
-              sandboxBackend?.internetAccess ?? "NONE"
-            )}
-          />
-          <CapabilityRow
-            label="dependencies"
-            value={getDependenciesLabel(
-              sandboxBackend?.supportsDependencies ?? false
-            )}
-          />
-        </Flex>
-      </Flex>
-    </ContextualHelp>
-  );
-}
-
-function MappingTile({
-  title,
-  description,
-  entries,
-  emptyLabel,
-  formatValue,
-}: {
-  title: string;
-  description: string;
-  entries: ReadonlyArray<[string, unknown]>;
-  emptyLabel: string;
-  formatValue: (value: unknown) => string;
-}) {
-  return (
-    <Flex direction="column" gap="size-100" elementType="section">
-      <Flex direction="column" gap="size-25">
-        <Text weight="heavy" size="S" elementType="h4">
-          {title}
-        </Text>
-        <Text size="XS" color="text-700">
-          {description}
-        </Text>
-      </Flex>
-      {entries.length === 0 ? (
-        <Text size="XS" color="text-500">
-          {emptyLabel}
-        </Text>
-      ) : (
-        <Flex direction="column" gap="size-75">
-          {entries.map(([key, value]) => (
-            <Flex
-              key={key}
-              direction="row"
-              gap="size-100"
-              alignItems="baseline"
-            >
-              <Text size="S" fontFamily="mono" color="text-700">
-                {key}
-              </Text>
-              <Text size="S" color="text-500" aria-hidden="true">
-                →
-              </Text>
-              <Text size="S" fontFamily="mono">
-                {formatValue(value)}
-              </Text>
-            </Flex>
-          ))}
-        </Flex>
-      )}
-    </Flex>
-  );
-}
-
-/** Values that should render in muted-italic (off / none) vs plain mono. */
-const MUTED_SETTING_VALUES = new Set(["off", "none"]);
-
-/** Setting keys whose values are comma-separated lists best shown one-per-line. */
-const LIST_SETTING_KEYS = new Set(["env_vars", "dependencies"]);
-
-function SettingValue({
-  settingKey,
-  value,
-}: {
-  settingKey: string;
-  value: string;
-}) {
-  const isMuted = MUTED_SETTING_VALUES.has(value);
-  if (LIST_SETTING_KEYS.has(settingKey) && !isMuted) {
-    const items = value.split(", ").filter((s) => s.length > 0);
-    return (
-      <Flex direction="column" alignItems="end" gap="size-25">
-        {items.map((item) => (
-          <Text key={item} size="S" fontFamily="mono">
-            {item}
-          </Text>
-        ))}
-      </Flex>
-    );
-  }
-  return (
-    <Text
-      size="S"
-      fontFamily="mono"
-      color={isMuted ? "text-500" : undefined}
-      fontStyle={isMuted ? "italic" : undefined}
-    >
-      {value}
-    </Text>
-  );
-}
-
-function formatPathMappingValue(value: unknown): string {
-  return typeof value === "string" ? value : JSON.stringify(value);
-}
-
-function formatLiteral(value: unknown): string {
-  if (value == null) return "null";
-  if (typeof value === "string") return JSON.stringify(value);
-  if (typeof value === "boolean" || typeof value === "number") {
-    return String(value);
-  }
-  return JSON.stringify(value);
-}
-
-function getInternetAccessLabel(
-  internetAccess: SandboxBackendInfo["internetAccess"]
-) {
-  switch (internetAccess) {
-    case "BOOLEAN":
-      return "Configurable";
-    case "NONE":
-      return "Not supported";
-    default:
-      return internetAccess;
-  }
-}
-
-function getDependenciesLabel(
-  supportsDependencies: SandboxBackendInfo["supportsDependencies"]
-) {
-  return supportsDependencies ? "Supported" : "Not supported";
-}
+import { LanguageWithIcon } from "@phoenix/pages/settings/sandboxes/utils";
 
 export function CodeDatasetEvaluatorDetails({
   datasetEvaluatorRef,
-  sandboxBackends,
 }: {
   datasetEvaluatorRef: CodeDatasetEvaluatorDetails_datasetEvaluator$key;
-  sandboxBackends: ReadonlyArray<SandboxBackendInfo>;
 }) {
   const datasetEvaluator = useFragment(
     graphql`
       fragment CodeDatasetEvaluatorDetails_datasetEvaluator on DatasetEvaluator {
-        id
         inputMapping {
           literalMapping
           pathMapping
@@ -260,9 +51,6 @@ export function CodeDatasetEvaluatorDetails({
         evaluator {
           kind
           ... on CodeEvaluator {
-            id
-            name
-            description
             language
             outputConfigs {
               __typename
@@ -287,25 +75,7 @@ export function CodeDatasetEvaluatorDetails({
               }
             }
             sandboxConfig {
-              id
-              name
-              description
-              timeout
-              config {
-                envVars {
-                  name
-                  secretKey
-                }
-                internetAccess {
-                  mode
-                }
-                dependencies {
-                  packages
-                }
-              }
-              provider {
-                backendType
-              }
+              ...CodeEvaluatorSandboxCard_sandboxConfig
             }
             currentVersion {
               sourceCode
@@ -327,52 +97,6 @@ export function CodeDatasetEvaluatorDetails({
     datasetEvaluator.outputConfigs.length > 0
       ? datasetEvaluator.outputConfigs
       : (evaluator.outputConfigs ?? []);
-  const sandboxConfig = evaluator.sandboxConfig ?? null;
-  const sandboxBackendByType = useMemo(
-    () =>
-      new Map(
-        sandboxBackends.map((sandboxBackend) => [
-          sandboxBackend.backendType,
-          sandboxBackend,
-        ])
-      ),
-    [sandboxBackends]
-  );
-  const sandboxBackend =
-    sandboxConfig != null
-      ? sandboxBackendByType.get(sandboxConfig.provider.backendType)
-      : undefined;
-
-  const pathMappingEntries = useMemo(
-    () =>
-      Object.entries(
-        (datasetEvaluator.inputMapping.pathMapping as Record<
-          string,
-          unknown
-        >) ?? {}
-      ),
-    [datasetEvaluator.inputMapping.pathMapping]
-  );
-  const literalMappingEntries = useMemo(
-    () =>
-      Object.entries(
-        (datasetEvaluator.inputMapping.literalMapping as Record<
-          string,
-          unknown
-        >) ?? {}
-      ),
-    [datasetEvaluator.inputMapping.literalMapping]
-  );
-
-  const canManageSandboxes = useViewerCanManageSandboxes();
-
-  const customSettings = useMemo(
-    () =>
-      sandboxConfig == null
-        ? []
-        : getSandboxConfigSettings(sandboxConfig.config),
-    [sandboxConfig]
-  );
 
   // currentVersion can be null (e.g. fixtures, backfills) — render an
   // empty state rather than throwing.
@@ -400,115 +124,18 @@ export function CodeDatasetEvaluatorDetails({
           </Card>
         </Flex>
         <Flex direction="column" gap="size-200" minWidth={0}>
-          <Card
-            title={
-              <Flex direction="row" gap="size-100" alignItems="center">
-                <Icon svg={<Icons.HardDrive />} />
-                <span>Sandbox</span>
-              </Flex>
-            }
-            extra={
-              canManageSandboxes ? (
-                <LinkButton
-                  size="S"
-                  to="/settings/sandboxes"
-                  aria-label="Configure sandboxes"
-                  leadingVisual={<Icon svg={<Icons.Settings />} />}
-                />
-              ) : undefined
-            }
-          >
-            {sandboxConfig == null ? (
-              <View padding="size-200">
-                <Text color="text-700">No sandbox configuration selected.</Text>
-              </View>
-            ) : (
-              <EvaluatorDetailList>
-                <EvaluatorDetailRow label="Config">
-                  <Text size="S" fontFamily="mono">
-                    {sandboxConfig.name}
-                  </Text>
-                </EvaluatorDetailRow>
-                {sandboxConfig.description ? (
-                  <EvaluatorDetailRow label="Description">
-                    {sandboxConfig.description}
-                  </EvaluatorDetailRow>
-                ) : null}
-                <EvaluatorDetailRow
-                  label="Provider"
-                  labelExtra={
-                    <ProviderCapabilitiesHelp sandboxBackend={sandboxBackend} />
-                  }
-                >
-                  <Flex direction="row" gap="size-100" alignItems="center">
-                    <SandboxProviderIcon
-                      backendType={sandboxConfig.provider.backendType}
-                      height={16}
-                    />
-                    <Text size="S">
-                      {sandboxBackend?.displayName ??
-                        sandboxConfig.provider.backendType}
-                    </Text>
-                  </Flex>
-                </EvaluatorDetailRow>
-                <EvaluatorDetailRow label="Timeout">
-                  {`${sandboxConfig.timeout} seconds`}
-                </EvaluatorDetailRow>
-                {customSettings.map((setting) => (
-                  <EvaluatorDetailRow key={setting.key} label={setting.label}>
-                    <SettingValue
-                      settingKey={setting.key}
-                      value={setting.value}
-                    />
-                  </EvaluatorDetailRow>
-                ))}
-              </EvaluatorDetailList>
-            )}
-          </Card>
-          <Card
-            title={
-              outputConfigs.length > 1
-                ? `Evaluator Annotations (${outputConfigs.length})`
-                : "Evaluator Annotation"
-            }
-          >
-            <View padding="size-200">
-              <Flex direction="column" gap="size-200">
-                {outputConfigs.map((config, idx) => {
-                  // The union's unknown arm ("%other") selects no fields, so
-                  // the block renders it as an all-defaults freeform config.
-                  const outputConfig = config as OutputConfig;
-                  return (
-                    <OutputConfigBlock
-                      key={outputConfig.name || idx}
-                      config={outputConfig}
-                      typename={config.__typename}
-                    />
-                  );
-                })}
-              </Flex>
-            </View>
-          </Card>
-          <Card title="Input Mapping">
-            <View padding="size-200">
-              <div css={mapGridCSS}>
-                <MappingTile
-                  title="Path mapping"
-                  description="Map function args to fields on the example"
-                  entries={pathMappingEntries}
-                  emptyLabel="No paths set"
-                  formatValue={formatPathMappingValue}
-                />
-                <MappingTile
-                  title="Literal mapping"
-                  description="Pass fixed literal values to function args"
-                  entries={literalMappingEntries}
-                  emptyLabel="No literals set"
-                  formatValue={formatLiteral}
-                />
-              </div>
-            </View>
-          </Card>
+          <CodeEvaluatorSandboxCard
+            sandboxConfigRef={evaluator.sandboxConfig}
+          />
+          <EvaluatorAnnotationsCard
+            outputConfigs={outputConfigs}
+            singularTitle="Evaluator Annotation"
+            pluralTitle="Evaluator Annotations"
+          />
+          <EvaluatorInputMappingCard
+            inputMapping={datasetEvaluator.inputMapping}
+            pathMappingDescription="Map function args to fields on the example"
+          />
         </Flex>
       </div>
     </div>
