+++ js/app/src/pages/project/evaluators/AnnotationConfigurationCard.tsx
@@ -1,19 +1,18 @@
 import { useFragment } from "react-relay";
 import { graphql } from "relay-runtime";
 
-import { Card, View } from "@phoenix/components";
-import { OutputConfigBlock } from "@phoenix/components/evaluators/OutputConfigBlock";
+import { EvaluatorAnnotationsCard } from "@phoenix/components/evaluators/EvaluatorAnnotationsCard";
 import { inferIncludeExplanationFromPrompt } from "@phoenix/components/evaluators/utils";
 import type { AnnotationConfigurationCard_projectEvaluator$key } from "@phoenix/pages/project/evaluators/__generated__/AnnotationConfigurationCard_projectEvaluator.graphql";
 
 /**
- * The annotation an LLM project evaluator writes: its name, which direction is
- * good, the values it can emit, and whether it explains itself.
+ * The annotations a project evaluator writes: their names, which direction is
+ * good, the values they can emit, and (for LLM evaluators) whether the
+ * evaluator explains itself.
  *
- * Split from the rest of the LLM configuration because it is a peer of Scope --
- * scope selects the work, this describes the result -- and the two stack in
- * the overview's aside. The grid inside matches the dataset evaluator pages,
- * so an annotation reads the same wherever it appears.
+ * Split from the rest of the evaluator configuration because it is a peer of
+ * Scope -- scope selects the work, this describes the result -- and the two
+ * stack in the overview's aside.
  */
 export function AnnotationConfigurationCard({
   projectEvaluatorRef,
@@ -24,6 +23,29 @@ export function AnnotationConfigurationCard({
     graphql`
       fragment AnnotationConfigurationCard_projectEvaluator on ProjectEvaluator {
         evaluator {
+          kind
+          outputConfigs {
+            __typename
+            ... on CategoricalAnnotationConfig {
+              name
+              optimizationDirection
+              values {
+                label
+                score
+              }
+            }
+            ... on ContinuousAnnotationConfig {
+              name
+              optimizationDirection
+              lowerBound
+              upperBound
+            }
+            ... on FreeformAnnotationConfig {
+              name
+              optimizationDirection
+              threshold
+            }
+          }
           ... on LLMEvaluator {
             promptVersion {
               tools {
@@ -37,28 +59,6 @@ export function AnnotationConfigurationCard({
                 }
               }
             }
-            outputConfigs {
-              __typename
-              ... on CategoricalAnnotationConfig {
-                name
-                optimizationDirection
-                values {
-                  label
-                  score
-                }
-              }
-              ... on ContinuousAnnotationConfig {
-                name
-                optimizationDirection
-                lowerBound
-                upperBound
-              }
-              ... on FreeformAnnotationConfig {
-                name
-                optimizationDirection
-                threshold
-              }
-            }
           }
         }
       }
@@ -66,25 +66,19 @@ export function AnnotationConfigurationCard({
     projectEvaluatorRef
   );
   const { evaluator } = projectEvaluator;
-  const outputConfig = evaluator.outputConfigs?.[0];
-  // An evaluator whose prompt has no explanation tool does not explain itself.
-  const includeExplanation = inferIncludeExplanationFromPrompt(
-    evaluator.promptVersion?.tools
-  );
-
-  if (outputConfig == null || outputConfig.__typename === "%other") {
-    return null;
-  }
+  // An LLM evaluator whose prompt has no explanation tool does not explain
+  // itself; code evaluators have no explanation toggle, so the cell is omitted.
+  const includeExplanation =
+    evaluator.kind === "LLM"
+      ? inferIncludeExplanationFromPrompt(evaluator.promptVersion?.tools)
+      : undefined;
 
   return (
-    <Card title="Annotation Configuration">
-      <View padding="size-200">
-        <OutputConfigBlock
-          config={outputConfig}
-          typename={outputConfig.__typename}
-          includeExplanation={includeExplanation}
-        />
-      </View>
-    </Card>
+    <EvaluatorAnnotationsCard
+      outputConfigs={evaluator.outputConfigs}
+      singularTitle="Annotation Configuration"
+      pluralTitle="Annotation Configurations"
+      includeExplanation={includeExplanation}
+    />
   );
 }
