+++ js/app/src/pages/project/evaluators/projectEvaluatorDetailsLoader.ts
@@ -15,8 +15,12 @@ export const projectEvaluatorDetailsLoaderGQL = graphql`
         name
         enabled
         evaluator {
-          kind
+          __typename
           description
+          ... on CodeEvaluator {
+            id
+            versionCount
+          }
         }
         traceProject {
           id
@@ -28,6 +32,7 @@ export const projectEvaluatorDetailsLoaderGQL = graphql`
         ...ProjectEvaluatorScopeDetails_projectEvaluator
         ...LLMProjectEvaluatorDetails_projectEvaluator
         ...CodeProjectEvaluatorDetails_projectEvaluator
+        ...CodeProjectEvaluatorConfigCards_projectEvaluator
         ...ProjectEvaluatorMetrics_projectEvaluator
         ...AnnotationConfigurationCard_projectEvaluator
       }
